add some log to locate ci failure
