+++ core/src/test/scala/org/apache/spark/sql/BaseTiSparkTest.scala
@@ -396,29 +396,6 @@ class BaseTiSparkTest extends QueryTest with SharedSQLContext {
     }
   }
 
-  private def listToString(result: List[List[Any]]): String =
-    if (result == null) s"[len: null] = null"
-    else if (result.isEmpty) s"[len: 0] = Empty"
-    else s"[len: ${result.length}] = ${result.map(mapStringList).mkString(",")}"
-
-  private def mapStringList(result: List[Any]): String =
-    if (result == null) "null" else "List(" + result.map(mapString).mkString(",") + ")"
-
-  private def mapString(result: Any): String =
-    if (result == null) "null"
-    else
-      result match {
-        case _: Array[Byte] =>
-          var str = "["
-          for (s <- result.asInstanceOf[Array[Byte]]) {
-            str += " " + s.toString
-          }
-          str += " ]"
-          str
-        case _ =>
-          result.toString
-      }
-
   protected def explainTestAndCollect(sql: String): Unit = {
     val df = spark.sql(sql)
     df.explain
